Take a custom command in batch mode
In batch mode radeco simply decompiles all the functions and outputs
the result. While this is nice, sometimes we want to run a specific
command.

Add a cli option to pass a custom command. Otherwise default to
decompile all (to mantain the old behaviour if `-c` is not passed).

Usage example:
    ./target/debug/radeco -b -c "dot entry0" tests/hello/hello > entry0.dot