Add some traits for calling analyzers
Right now there is not a standard way to implement/call an analyzer. For
example CSE builds a struct, while Inferer provides the `run` function.

First of, all the analyzers should provide a name and their kind as requested by
the `Analyzer` trait.
Then, if an an analyszer works on an entire `RadecoModule` it should implement
the `ModuleAnalyzer` trait, if it works on a single `RadecoFunction` it should
implement the `FuncAnalyzer` trait.

They can also return an `AnalyzerResult` to report extra infos.