Use a policy to selectively commit changes to the IR
Sometimes applying all the changes an `Analyzer` produces is not the
wanted behaviour. Tweak the `analyze` API to require a policy function
which is responsible of applying or discarding the produced changes. The
policy is called on the fly by the `Analyzer` as soon as a new change is
made to decide whether to apply it or to discard it.

As changes are applied, the IR is not the same as before, thus changes
previously discarded could be proposed again by the `Analyer`. On the
other hand an `Analyzer` is not expected to propose again a change if
all the previous other changes were skipped.

Since not all the `Analyzer`s are "change-oriented" a `uses_policy`
api is added so that `Analyzer`s can report whether or not will they use
the policy function.