Use an `Engine` to run analysis passes
Currently, analyzers are run by radeco-lib's consumers. This could be
useful in some cases, but in general it is desirable to implement this
logic inside radeco-lib.

Add a `Engine` trait which is responsible of running the available
analyzers in order to simplify the IR.

For now, implement that trait for `RadecoEngine`. In the future it would
be possible to add more engines, each of them tuned to achieve different
results.