Add the attribute `kind` to `RadecoFunction`
We cannot load the code of imported or relocated functions, because it
is not available inside the binary file.

Fixes #254. Previously, the fact that some functions were relocated was
ignored and thus some garbage code was imported (r2 returns some garbage
code when `pdf` is called on relocated functions). Unfortunately that
codes generates a huge IR graph, and this leads to a crash when the matcher
is called (memory exhaustion due to veeery long strings).